chore: rename displays to utils
