+++ src/inputs/main_option.c
@@ -1,4 +1,4 @@
-#include "../displays/ansi_color.h"
+#include "../utils/ansi_color.h"
 #include <stdio.h>
 #include <stdlib.h>
