+++ src/number_systems/binary_system/binary_description.c
@@ -1,6 +1,6 @@
-#include "../../displays/ansi_color.h"
-#include "../../displays/custom_print.h"
-#include "../../displays/escape_sequence.h"
+#include "../../utils/ansi_color.h"
+#include "../../utils/custom_print.h"
+#include "../../utils/escape_sequence.h"
 
 void binary_description(void) {
   text_blue("--------------------------------------------------------", 15);
