+++ src/number_systems/binary_system/binary_system.c
@@ -1,6 +1,6 @@
-#include "../../displays/ansi_color.h"
-#include "../../displays/clear_terminal.h"
-#include "../../displays/escape_sequence.h"
+#include "../../utils/ansi_color.h"
+#include "../../utils/clear_terminal.h"
+#include "../../utils/escape_sequence.h"
 #include "binary_description.h"
 #include "binary_to_decimal.h"
 #include "binary_validate.h"
