+++ src/number_systems/binary_system/binary_to_decimal.c
@@ -1,4 +1,4 @@
-#include "../../displays/ansi_color.h"
+#include "../../utils/ansi_color.h"
 #include "reverse_string.h"
 #include <math.h>
 #include <stdlib.h>
