+++ src/number_systems/binary_system/reverse_string.c
@@ -1,12 +1,11 @@
-#include <string.h>
 #include <stdlib.h>
-#include <stdio.h>
+#include <string.h>
 
-char* reverse_string(char *input) {
+char *reverse_string(char *input) {
   int input_length = strlen(input);
   int current_index = input_length - 1;
   int index_increment = 0;
-  char* reversed_string = malloc(sizeof(char) * input_length);
+  char *reversed_string = malloc(sizeof(char) * input_length);
 
   for (; current_index > -1; current_index--) {
     reversed_string[index_increment] = input[current_index];
