+++ src/utils/app_description.c
@@ -1,5 +1,5 @@
-#include "../displays/custom_print.h"
 #include "ansi_color.h"
+#include "custom_print.h"
 #include "escape_sequence.h"
 
 void app_description(void) {
