+++ src/utils/clear_terminal.c
