+++ src/utils/clear_terminal.h
